fix: separate in pages and components
